fix: faulty logic update causes script to always display usage information (#160)
Updates for linting changed the logic from `[ "$checksum_file" = ""]`
which is true when there is no filename as the first argument to `[[ -n
"${checksum_file}" ]]` which is true when there **is** a filename as the
first argument. This is the reverse of the original logic and causes the
script to always fail.

I changed the logic to be a test on the number of arguments instead of
testing the length of a string in arg position 1. This will catch
additional errors where callers may attempt to call this with extra
arguments thinking that it can support multiple checksum files.